requirements_mediapipe.txt
pip==23.2.1
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.functools-lru-cache==1.6.5
beautifulsoup4==4.12.2
bleach==6.0.0
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
flatbuffers==2.0
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.12.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.3.1
jupyter-console==6.6.3
jupyter_core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
jupyterlab-widgets==3.0.9
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mediapipe==0.10.5
mistune==3.0.1
multidict==6.0.4
munkres==1.1.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.6
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.23.2
oauthlib==3.2.2
openai==0.28.1
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandas-stubs==1.2.0.62
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.17.0
pooch==1.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.31.0
setproctitle==1.3.2
setuptools==68.2.2
six==1.15.0
smmap==3.0.5
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
stack-data==0.6.2
statsmodels==0.14.0
tenacity==8.2.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.8.0
typing-utils==0.1.0
tzdata==2023.3
unicodedata2==15.0.0
uri-template==1.3.0
urllib3==1.26.16
wandb==0.15.11
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.7
wheel==0.41.2
widgetsnbextension==4.0.9
wrapt==1.15.0
yarl==1.9.2
uvicorn
fastapi
opencv-python
python-multipart
starlette